@@ -5,7 +5,7 @@ author: Edouard Marquez <moi@edouard-marquez.me>
 homepage: https://github.com/g123k/flutter_fimber_filelogger
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
